Refactored lexer_get_size.c
